test(textarea): migrate autogrow
