+++ core/src/components/textarea/test/autogrow/textarea.e2e.ts-snapshots/textarea-autogrow-diff-md-ltr-Mobile-Safari-linux.png
